requirements.txt
pandas
numpy
xgboost
matplotlib
scikit-learn
joblib
shap
fastapi
python-dotenv
fastapi
uvicorn
pydantic
uvicorn
